test(encoding): split the checked-read file along the line the runners draw
`checkedReadMigration.test.ts` was two files sharing a name, which is why
neither answer to "should this move?" fit it. #635 read it as mostly
never-migrate and left the whole thing behind; a rename would have
carried the never-migrate half across. The split is per test.

Six stayed. One asserts that the unchecked `read_file_content` is called
nowhere in `src` and is gone from the Rust crate — an absence, which no
run can observe (executing the checked read cannot prove there is no
other caller), and then the same claim across a language boundary where
nothing but the spelling connects the two sides. The other five pin the
ORDER of statements inside `.svelte` handlers: flag the tab before the
buffer is published, recognise the refusal before raising the generic
toast. Mounting the component to earn a weaker version of that would put
jsdom in for Monaco, mermaid and KaTeX.

Three moved. They drive `ensureFullContent` through the real store and
the real session, both runes modules, and under `node --test` they only
ran because this file installed `$state`/`$derived`/`$effect` onto
`globalThis`. That shim stood exactly where the assertions look —
`tab.isTruncated` and `tab.hasReplacementChars` are read straight back
out of the store. They are `checkedReadMigration.spec.ts` now, built by
the Svelte compiler.

Dropping `setTabDecodedLossy` from `ensureFullContent` fails all three of
the moved tests; putting an `invoke('read_file_content')` back into
`documentSession.svelte.ts` fails the absence test that stayed. Both
halves are live, on their own side.
